requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
deprecation==2.1.0
exceptiongroup==1.3.0
fastapi==0.116.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
packaging==25.0
postgrest==2.19.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
realtime==2.19.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
storage3==2.19.0
StrEnum==0.4.15
supabase==2.19.0
supabase-auth==2.19.0
supabase-functions==2.19.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.36.0
websockets==15.0.1
